% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addIntercept}
\alias{addIntercept}
\title{addIntercept add intercept column to design matrix}
\usage{
addIntercept(x, check = FALSE)
}
\arguments{
\item{check}{}
}
\value{

}
\description{
addIntercept add intercept column to design matrix
}
