% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.RobZS.R
\name{coef.RobZS}
\alias{coef.RobZS}
\title{coef.RobZS}
\usage{
\method{coef}{RobZS}(object, vers = c("reweighted", "raw"), zeros = TRUE, ...)
}
\arguments{
\item{object}{fitted RobZS model object.}

\item{vers}{a character string specifying for which fit to make predictions. Possible values are reweighted (the default) for predicting values from the reweighted fit, raw for predicting values from the raw fit.}

\item{zeros}{a logical indicating whether to keep zero coefficients ('TRUE', the default) or to omit them ('FALSE').}

\item{...}{additional arguments from the RobZS object if needed.}
}
\value{
a numeric vector containing the requested coefficients.
}
\description{
coef.RobZS
}
