% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.RobZS.R
\name{cv.RobZS}
\alias{cv.RobZS}
\title{cv.RobZS}
\usage{
cv.RobZS(
  index = NULL,
  xx,
  yy,
  family,
  h,
  alphas,
  lambdas,
  nfold,
  repl,
  ncores,
  plot = TRUE
)
}
\arguments{
\item{plot}{}
}
\value{

}
\description{
cv.RobZS
}
