% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.RobZS.R
\name{fitted.RobZS}
\alias{fitted.RobZS}
\title{fitted.RobZS}
\usage{
\method{fitted}{RobZS}(
  object,
  vers = c("reweighted", "raw", "both"),
  type0 = c("response"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
fitted.RobZS
}
