% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RobZS.R
\name{plot.RobZS}
\alias{plot.RobZS}
\title{plot.RobZS}
\usage{
\method{plot}{RobZS}(
  x,
  method = c("coefficients", "resid", "diagnostic"),
  vers = c("reweighted", "raw"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot.RobZS
}
