% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoef.RobZS.R
\name{plotCoef.RobZS}
\alias{plotCoef.RobZS}
\title{plotCoef.RobZS}
\usage{
plotCoef.RobZS(object, vers = c("reweighted", "raw"), colors = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plotCoef.RobZS
}
