% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostic.RobZS.R
\name{plotDiagnostic.RobZS}
\alias{plotDiagnostic.RobZS}
\title{plotDiagnostic.RobZS}
\usage{
plotDiagnostic.RobZS(object, vers = c("reweighted", "raw"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plotDiagnostic.RobZS
}
