% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResid.RobZS.R
\name{plotResid.RobZS}
\alias{plotResid.RobZS}
\title{plotResid.RobZS}
\usage{
plotResid.RobZS(object, vers = c("reweighted", "raw"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plotResid.RobZS
}
