% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.RobZS.R
\name{predict.RobZS}
\alias{predict.RobZS}
\title{predict.RobZS}
\usage{
\method{predict}{RobZS}(
  object,
  newX,
  vers = c("reweighted", "raw", "both"),
  type0 = c("response", "coefficients", "nonzero"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
predict.RobZS
}
