% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prepara0}
\alias{prepara0}
\title{prepara0}
\usage{
prepara0(x, y, family = "gaussian", index = NULL, robu = NULL)
}
\arguments{
\item{x}{data.frame the covariate matrix}

\item{y}{numeric the response vector}

\item{family}{character gaussian}

\item{index}{numeric}

\item{robu}{integer}
}
\value{
list
}
\description{
prepara0
}
