% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.RobZS.R
\name{print.RobZS}
\alias{print.RobZS}
\title{print.RobZS}
\usage{
\method{print}{RobZS}(x, vers = c("reweighted", "raw"), zeros = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
print.RobZS
}
