% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.RobZS.R
\name{residuals.RobZS}
\alias{residuals.RobZS}
\title{residuals.RobZS}
\usage{
\method{residuals}{RobZS}(object, vers = c("reweighted", "raw", "both"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
residuals.RobZS
}
