% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobZS.R
\name{RobZS}
\alias{RobZS}
\title{RobZS}
\usage{
RobZS(
  xx,
  yy,
  family = c("gaussian"),
  alphas,
  lambdas,
  lambdaw,
  hsize = 0.75,
  intercept = TRUE,
  nsamp = 500,
  s1 = 10,
  nCsteps = 20,
  nfold = 5,
  seed = NULL,
  plot = TRUE,
  repl = 5,
  para = FALSE,
  ncores = 1,
  del = 0.0125,
  tol = -1e+06,
  type0 = c("response")
)
}
\arguments{
\item{type0}{}
}
\value{

}
\description{
RobZS
}
