% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.RobZS.R
\name{weights.RobZS}
\alias{weights.RobZS}
\title{weights.RobZS}
\usage{
\method{weights}{RobZS}(object, vers = c("reweighted", "raw", "both"), index = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
weights.RobZS
}
